How about we just hook into the existing auto advance feature instead and remove the need for any special organizing of the cards into various decks?

+++ src/MagnusAddon/note/difficulty_calculator.py
@@ -1,13 +1,13 @@
 from sysutils import ex_sequence, kana_utils
 
-KANJI_WEIGHT = 4
+KANJI_WEIGHT = 4.0
 KATAKANA_WEIGHT = 1.5
 
-def find_difficulty(string: str) -> int:
+def find_difficulty(string: str) -> float:
     kanji_characters = ex_sequence.count(string, kana_utils.is_kanji)
     katakana_characters = ex_sequence.count(string, kana_utils.is_katakana)
     hiragana_characters = ex_sequence.count(string, kana_utils.is_hiragana)
-    return int(kanji_characters * KANJI_WEIGHT + katakana_characters * KATAKANA_WEIGHT + hiragana_characters)
+    return kanji_characters * KANJI_WEIGHT + katakana_characters * KATAKANA_WEIGHT + hiragana_characters
 
 class DifficultyCalculator:
     def __init__(self, levels: list[int], difficulties: list[float]) -> None:
